requirements.txt
greenlet==3.3.0
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
psycopg==3.3.2
pytesseract==0.3.13
python-dotenv==1.2.1
SQLAlchemy==2.0.45
typing_extensions==4.15.0
